+++ projects/igniteui-angular/src/lib/drop-down/drop-down.component.ts
@@ -192,7 +192,7 @@ export class IgxDropDownBase implements OnInit, IToggleView {
      * ```
      */
     set id(value: string) {
-        this.selectionAPI.set_selection(value, this.selectionAPI.get_selection(this.id));
+        this.selection.set(value, this.selection.get(this.id));
         this._id = value;
         this.toggleDirective.id = value;
     }
@@ -216,13 +216,12 @@ export class IgxDropDownBase implements OnInit, IToggleView {
      * ```
      */
     public get selectedItem(): any {
-        const selectedItem = this.selectionAPI.get_selection_first(this.id);
-
+        const selectedItem = this.selection.first_item(this.id);
         if (selectedItem) {
             if (selectedItem.isSelected) {
                 return selectedItem;
             }
-            this.selectionAPI.set_selection(this.id, new Set());
+            this.selection.clear(this.id);
         }
         return null;
     }
@@ -288,7 +287,7 @@ export class IgxDropDownBase implements OnInit, IToggleView {
     constructor(
         protected elementRef: ElementRef,
         protected cdr: ChangeDetectorRef,
-        protected selectionAPI: IgxSelectionAPIService) { }
+        protected selection: IgxSelectionAPIService) { }
 
     /**
      * Select an item by index
@@ -403,7 +402,7 @@ export class IgxDropDownBase implements OnInit, IToggleView {
      */
     ngOnInit() {
         this.toggleDirective.id = this.id;
-        this.selectionAPI.set_selection(this.id, new Set());
+        this.selection.clear(this.id);
     }
 
 
@@ -497,7 +496,7 @@ export class IgxDropDownBase implements OnInit, IToggleView {
         }
 
         const args: ISelectionEventArgs = { oldSelection, newSelection };
-        this.selectionAPI.set_selection(this.id, new Set([newSelection]));
+        this.selection.set(this.id, new Set([newSelection]));
 
         this.onSelection.emit(args);
     }
@@ -689,8 +688,8 @@ export class IgxDropDownComponent extends IgxDropDownBase {
     constructor(
         protected elementRef: ElementRef,
         protected cdr: ChangeDetectorRef,
-        protected selectionAPI: IgxSelectionAPIService) {
-        super(elementRef, cdr, selectionAPI);
+        protected selection: IgxSelectionAPIService) {
+        super(elementRef, cdr, selection);
     }
 
     protected changeSelectedItem(newSelection?: IgxDropDownItemComponent) {